Fix: Remove lazy loading on feedback button  (#2963)
